yaml options support humanize

+++ aoss/stream-store/wal/reload.go
@@ -114,7 +114,7 @@ func Reload(options Options, fn func(streamstorepb.Entry) error) (*wal, error) {
 		}
 
 		//close files
-		if i < len(fileInfos)-1 || fileInfo.Size >= options.MaxLogSize {
+		if i < len(fileInfos)-1 || fileInfo.Size >= int64(options.MaxLogSize) {
 			lf, err := initLogFile(fileInfo.f, fileInfo.firstID, fileInfo.lastID)
 			if err != nil {
 				return nil, errors.WithStack(err)
